Cambios en producto

+++ src/main/java/com/TPfinal/Productos/model/entity/Alimento.java
@@ -3,7 +3,7 @@ package com.TPfinal.Productos.model.entity;
 public class Alimento extends Producto{
     private String vencimiento;
 
-    public Alimento(String nombre, Integer precio, String marca, Integer stockDisponible, String vencimiento) {
+    public Alimento(String nombre, Double precio, String marca, Integer stockDisponible, String vencimiento) {
         super(nombre, precio, marca, stockDisponible);
         this.vencimiento = vencimiento;
     }
